Add representative search form
Add a form that allows a user to look up a represenative by address
and filter the timeline view to only that rep.

It is initialized with callbacks that pass data to and from calling
code.

In the app class, use the Google Civic Information API to lookup the
U.S. Representative for the address.