Super-basic "loading" message
In production, the data is loaded from an API deployed on a free-tier
Heroku dyno.  This can cause a noticeable lag in the amount of time it
takes for the timeline to render.

Add a very basic loading message just so users know to expect something
to happen in this space.  For now, it's just text that gets blown away
when the timelne renders.  In the future it could be some kind of
animation.

Addresses #8