Render timeline to a string
Add a script to render the timeline to a string.

This makes it easy to copy and paste initial HTML into WordPress.

In order to make it possible to render the timeline in a separate
script, factor some of the code out of the `App` class.

Add `jsdom` and `babel-require` as dependencies as these are needed
by the new script.

Addresses #9